massive refactor - now a module
