--- PythonScripts/UniFolderCreator/requirements.txt
@@ -1,1 +1,0 @@
-typer==0.6.1
