+++ PythonScripts/UniFolderCreatorProject/UniFolderCreator/__main__.py
@@ -1,0 +1,7 @@
+from UniFolderCreator import cli, __app_name__
+
+def main():
+    cli.app(prog_name=__app_name__)
+
+if __name__ == "__main__":
+    main()
