+++ PythonScripts/UniFolderCreatorProject/UniFolderCreator/tree.py
@@ -1,0 +1,32 @@
+import os
+
+# Credit: https://stackoverflow.com/a/45599996
+# Prints a directory tree after running
+def realname(path, root=None):
+    if root is not None:
+        path=os.path.join(root, path)
+    result=os.path.basename(path)
+    if os.path.islink(path):
+        realpath=os.readlink(path)
+        result= '%s -> %s' % (os.path.basename(path), realpath)
+    return result
+
+def ptree(startpath, depth=-1):
+    prefix=0
+    if startpath != '/':
+        if startpath.endswith('/'): startpath=startpath[:-1]
+        prefix=len(startpath)
+    for root, dirs, files in os.walk(startpath):
+        level = root[prefix:].count(os.sep)
+        if depth >-1 and level > depth: continue
+        indent=subindent =''
+        if level > 0:
+            indent = '|   ' * (level-1) + '|-- '
+        subindent = '|   ' * (level) + '|-- '
+        print('{}{}/'.format(indent, realname(root)))
+        # print dir only if symbolic link; otherwise, will be printed as root
+        for d in dirs:
+            if os.path.islink(os.path.join(root, d)):
+                print('{}{}'.format(subindent, realname(d, root=root)))
+        for f in files:
+            print('{}{}'.format(subindent, realname(f, root=root)))
