+++ PythonScripts/UniFolderCreatorProject/requirements.txt
@@ -1,0 +1,4 @@
+typer==0.6.1
+colorama==0.4.5
+termcolor==1.1.0
+pytest==7.1.2
