+++ PythonScripts/UniFolderCreatorProject/tests/test_udirc.py
@@ -1,0 +1,9 @@
+from typer.testing import CliRunner
+from UniFolderCreator import __app_name__, __version__, cli
+
+runner = CliRunner()
+
+def test_version():
+    result = runner.invoke(cli.app, ["--version"])
+    assert result.exit_code == 0
+    assert f"{__app_name__} v{__version__}\n" in result.stdout
